wrapper script to replace dnsmasq.conf placeholders
this script should run as a daemon when installed via the snap.  It
will replace the values in dnsmasq.conf for those appropriate to
Pi-hole as well as the system it is installed on.

Currently, the script assumes the interface and IP addresses to use.
Since this is still alpha software, I'm just going with something I
know will work to get an appropriate value.